requirements.txt
numpy
pandas
pillow
matplotlib
scipy
scikit-learn
torch
torchvision
google-search-results
requests
